CM-104 Modificar users-be para que opere segun los permisos del usuar…

+++ src/main/java/com/cema/users/controllers/UserController.java
@@ -1,26 +1,26 @@
 package com.cema.users.controllers;
 
 import com.cema.users.constants.Messages;
+import com.cema.users.constants.Roles;
 import com.cema.users.domain.User;
 import com.cema.users.entities.CemaUser;
+import com.cema.users.exceptions.UnauthorizedException;
 import com.cema.users.exceptions.UserExistsException;
 import com.cema.users.exceptions.UserNotFoundException;
 import com.cema.users.mapping.UserMapping;
 import com.cema.users.repositories.CemaUserRepository;
-import com.cema.users.services.login.LoginService;
+import com.cema.users.services.authorization.AuthorizationService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import io.swagger.annotations.ResponseHeader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -39,26 +39,27 @@ import java.util.stream.Collectors;
 @RequestMapping("/v1")
 @Api(produces = "application/json", value = "Allows interaction with the users database and authorization operations. V1")
 @Validated
-public class Controller {
+public class UserController {
 
     private static final String BASE_URL = "/users/";
 
-    private static final Logger LOG = LoggerFactory.getLogger(Controller.class);
+    private static final Logger LOG = LoggerFactory.getLogger(UserController.class);
 
     private final CemaUserRepository cemaUserRepository;
     private final UserMapping userMapping;
-    private final LoginService loginService;
+    private final AuthorizationService authorizationService;
 
-    public Controller(CemaUserRepository cemaUserRepository, UserMapping userMapping, LoginService loginService) {
+    public UserController(CemaUserRepository cemaUserRepository, UserMapping userMapping, AuthorizationService authorizationService) {
         this.cemaUserRepository = cemaUserRepository;
         this.userMapping = userMapping;
-        this.loginService = loginService;
+        this.authorizationService = authorizationService;
     }
 
     @ApiOperation(value = "Retrieve a user data by username", response = User.class)
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Successfully found user"),
-            @ApiResponse(code = 404, message = "The user you were looking for is not found")
+            @ApiResponse(code = 404, message = "The user you were looking for is not found"),
+            @ApiResponse(code = 401, message = "You are not allowed to look for this user")
     })
     @GetMapping(value = BASE_URL + "{username}", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<User> lookUpUser(
@@ -73,15 +74,20 @@ public class Controller {
         if (cemaUser == null) {
             throw new UserNotFoundException(String.format(Messages.USER_DOES_NOT_EXISTS, userName));
         }
+        if (!authorizationService.isOnTheSameEstablishment(cemaUser.getEstablishmentCuig())) {
+            throw new UnauthorizedException(String.format(Messages.OUTSIDE_ESTABLISHMENT, cemaUser.getEstablishmentCuig()));
+        }
         User user = userMapping.mapEntityToDomain(cemaUser);
 
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
+    @PreAuthorize("hasRole('PATRON')")
     @ApiOperation(value = "Register a new user to the database")
     @ApiResponses(value = {
             @ApiResponse(code = 201, message = "User created successfully"),
-            @ApiResponse(code = 409, message = "The user you were trying to create already exists")
+            @ApiResponse(code = 409, message = "The user you were trying to create already exists"),
+            @ApiResponse(code = 401, message = "You are not allowed to register this user")
     })
     @PostMapping(value = BASE_URL + "register", produces = {MediaType.APPLICATION_JSON_VALUE}, consumes = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<User> register(
@@ -96,12 +102,18 @@ public class Controller {
         user.setUserName(userName);
         LOG.info("Request to register user: {}", userName);
 
+        if (!authorizationService.isOnTheSameEstablishment(user.getEstablishmentCuig())) {
+            throw new UnauthorizedException(String.format(Messages.OUTSIDE_ESTABLISHMENT, user.getEstablishmentCuig()));
+        }
+        if(!authorizationService.isAdmin() && user.getRole().equalsIgnoreCase(Roles.ADMIN)){
+            throw new UnauthorizedException(String.format(Messages.ACTION_NOT_ALLOWED, user.getRole()));
+        }
+
         CemaUser cemaUser = cemaUserRepository.findCemaUserByUserName(userName);
         if (cemaUser != null) {
             LOG.info("User already exists");
             throw new UserExistsException(String.format(Messages.USER_ALREADY_EXISTS, userName));
         }
-
         cemaUser = userMapping.mapDomainToEntity(user, userName, password);
 
         cemaUserRepository.save(cemaUser);
@@ -109,10 +121,12 @@ public class Controller {
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
+    @PreAuthorize("hasRole('PATRON')")
     @ApiOperation(value = "Delete an existing user by username")
     @ApiResponses(value = {
             @ApiResponse(code = 201, message = "User deleted successfully"),
-            @ApiResponse(code = 404, message = "The user you were trying to reach is not found")
+            @ApiResponse(code = 404, message = "The user you were trying to reach is not found"),
+            @ApiResponse(code = 401, message = "You are not allowed to delete this user")
     })
     @DeleteMapping(value = BASE_URL + "{username}", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<User> delete(
@@ -125,6 +139,9 @@ public class Controller {
         userName = userName.toLowerCase();
         CemaUser cemaUser = cemaUserRepository.findCemaUserByUserName(userName);
         if (cemaUser != null) {
+            if (!authorizationService.isOnTheSameEstablishment(cemaUser.getEstablishmentCuig())) {
+                throw new UnauthorizedException(String.format(Messages.OUTSIDE_ESTABLISHMENT, cemaUser.getEstablishmentCuig()));
+            }
             LOG.info("User exists, deleting");
             cemaUserRepository.delete(cemaUser);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -133,6 +150,7 @@ public class Controller {
         throw new UserNotFoundException(String.format(Messages.USER_DOES_NOT_EXISTS, userName));
     }
 
+    @PreAuthorize("hasRole('PATRON')")
     @ApiOperation(value = "Retrieve a list of users with the specified role", response = User.class, responseContainer = "List")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Successfully found user")
@@ -146,11 +164,20 @@ public class Controller {
 
         LOG.info("Searching users with role {}", role);
 
-        List<CemaUser> bovineList = cemaUserRepository.findCemaUsersByRole(role);
+        List<CemaUser> bovineList = cemaUserRepository.findCemaUsersByRoleIgnoreCase(role);
         LOG.info("Returned {} users from db", bovineList.size());
+        String currentCuig = authorizationService.getCurrentUserCuig();
 
-        List<User> mappedUsers = bovineList.stream().map(userMapping::mapEntityToDomain).collect(Collectors.toList());
+        List<User> mappedUsers = bovineList.stream()
+                .map(userMapping::mapEntityToDomain)
+                .collect(Collectors.toList());
 
+        //If the user is not admin we filter out external users
+        if(!authorizationService.isAdmin()){
+            mappedUsers = mappedUsers.stream()
+                    .filter(user -> user.getEstablishmentCuig().equals(currentCuig))
+                    .collect(Collectors.toList());
+        }
 
         return ResponseEntity.ok().body(mappedUsers);
     }
